square totally works

+++ app/controllers/application_controller.rb
@@ -5,7 +5,7 @@ class ApplicationController < ActionController::Base
   end
 
   def calc_square
-    @num = params.fetch("user_input_sq").to_i
+    @num = params.fetch("user_input_sq").to_f
     @sq_num = @num ** 2
 
     render({:template => "calc_temp/square_result.html.erb"})
